job-recommendation-system/requirements.txt
openai==0.28.0
psycopg2-binary==2.9.10
pgvector==0.3.6
loguru==0.7.2
python-jobspy==1.1.75
python-dotenv==1.0.1
pymupdf4llm==0.0.17